Force GPG key id to be short during rpm check
Previously there was an implicit expectation that gpg was using the short ID of
the key when outputing. However, this can be changed and often is because
seriously, shortids have conflicts. So, this forces it back to short for
purposes of verification in this module so you don't end up importing a new GPG
key on each puppet run.

Fixes #33